Fix some logic for CRUD

+++ controllers/categoryController.js
@@ -1,5 +1,6 @@
 const mongoose = require("mongoose");
 const Category = require("../models/category");
+const Product = require("../models/product/product");
 const ErrorResponse = require("../utils/errorResponse");
 
 exports.getAllCategories = async (req, res, next) => {
@@ -62,16 +63,24 @@ exports.createCategory = async (req, res, next) => {
     const { categoryName, categorySlug } = req.body;
 
     try {
-        const category = await Category.create({
-            categoryName,
-            categorySlug
+        const slug = await Category.findOne({
+            categorySlug: categorySlug
         });
 
-        res.status(201).json({
-            success: true,
-            message: "Category created successfully",
-            data: category
-        });
+        if (slug)
+            return next(new ErrorResponse("Category slug existed in database", 400));
+        else {
+            const category = await Category.create({
+                categoryName,
+                categorySlug
+            });
+    
+            res.status(201).json({
+                success: true,
+                message: "Category created successfully",
+                data: category
+            });
+        }
     } catch (error) {
         next(error);
     }
@@ -86,23 +95,33 @@ exports.updateCategory = async (req, res, next) => {
     const { categoryName, categorySlug } = req.body;
 
     try {
-        const category = await Category.findByIdAndUpdate(
-            categoryId,
-            { 
-                categoryName,
-                categorySlug 
-            },
-            { new: true }
-        );
-
-        if (!category)
-            return next(new ErrorResponse("No category found", 404));
-        
-        res.status(200).json({
-            success: true,
-            message: "Category updated successfully",
-            data: category
+        const slug = await Category.findOne({
+            _id: { $ne: categoryId },
+            categorySlug: categorySlug
         });
+
+        if (slug) {
+            return next(new ErrorResponse("Category slug existed in database", 400));
+        }
+        else {
+            const category = await Category.findByIdAndUpdate(
+                categoryId,
+                { 
+                    categoryName,
+                    categorySlug 
+                },
+                { new: true }
+            );
+    
+            if (!category)
+                return next(new ErrorResponse("No category found", 404));
+            
+            res.status(200).json({
+                success: true,
+                message: "Category updated successfully",
+                data: category
+            });
+        }
     } catch (error) {
         next(error);
     }
@@ -115,16 +134,25 @@ exports.deleteCategory = async (req, res, next) => {
         return next(new ErrorResponse("Please provide valid category's ID", 400));
 
     try {
-        const category = await Category.findByIdAndDelete(categoryId);
-
-        if (!category)
-            return next(new ErrorResponse("No category found", 404));
-        
-        res.status(200).json({
-            success: true,
-            message: "Category deleted successfully",
-            data: category
+        const prodsList = await Product.find({
+            productCategoryId: categoryId
         });
+
+        if (prodsList.length > 0) {
+            return next(new ErrorResponse("This category can't be deleted", 400));
+        }
+        else {
+            const category = await Category.findByIdAndDelete(categoryId);
+
+            if (!category)
+                return next(new ErrorResponse("No category found", 404));
+            
+            res.status(200).json({
+                success: true,
+                message: "Category deleted successfully",
+                data: category
+            });
+        }
     } catch (error) {
         next(error);
     }
